change grade requests

+++ bootcampspot/src/main/java/com/trilogy/bootcampspot/data/net/response/Response.kt
@@ -91,6 +91,6 @@ data class StudentResume (
     val maxAllowedAbsences: Int?,
     val remainingAbsenceCount: Int?,
     val careerCompletionPercentage: Int?,
-    val academicAverageValue: Int?,
+    val academicAverageValue: Float?,
     val academicAverageGrade: String?
 )
